Got Strikethrough Working (mostly)
I've got strikethrough working the same way that it was for the XML version, mostly, except only with whole lines, since we're kind of adding entire lines at once right now. For some reason, it often doesn't work unless you scroll over the middlemost part of a line, and sometimes doesn't work at all for especially small words. I had to make all lines buttons, which is what it was in the XML implementation.